Cleanup a little after multi project refactor
When I was going through the PR for the last change I made I found some superficial issues I wanted to fix. Extra white space, no white space, and renaming a package to be consistent with the other sub projects.
