+++ mpd/mpd.gradle
@@ -1,3 +1,1 @@
 apply from: "../base-extension.gradle"
-
-
