+++ mpd/src/main/kotlin/offthecob/mpd/TrackHandler.kt
@@ -1,4 +1,4 @@
-package com.offthecob
+package offthecob.mpd
 
 import com.bitwig.extension.controller.api.*
 import javax.inject.Inject
