+++ mpd/src/main/kotlin/offthecob/mpd/cc/ControlChange.kt
@@ -1,4 +1,4 @@
-package com.offthecob.cc
+package offthecob.mpd.cc
 
 interface ControlChange {
     fun cc(note: Int, value: Int)
