+++ mpd/src/main/kotlin/offthecob/mpd/cc/DeviceCc.kt
@@ -1,9 +1,9 @@
-package com.offthecob.cc
+package offthecob.mpd.cc
 
 import com.bitwig.extension.controller.api.CursorRemoteControlsPage
 import com.bitwig.extension.controller.api.PinnableCursorDevice
 import com.google.inject.Inject
-import com.offthecob.*
+import offthecob.mpd.*
 
 class DeviceCc @Inject constructor(private val trackHandler: TrackHandler) : ControlChange {
     override fun cc(note: Int, value: Int) {
