+++ mpd/src/main/resources/META-INF/services/com.bitwig.extension.ExtensionDefinition
@@ -1,1 +1,1 @@
-com.offthecob.MpdExtensionDefinition
+offthecob.mpd.MpdExtensionDefinition
